+++ docs/src/release_notes.rst
@@ -3,6 +3,13 @@
 
 Release notes
 =============
+Unreleased
+----------
+Added
+*****
+- :ref:`config` - warning messages for missing configuration now include the
+  variable name which was missing (:issue:`222`)
+
 3.2.0 (2020-10-16)
 ------------------
 Added